Move withQueryIntercept function to fixtures
